Movement Update
no more move on ice

+++ Project Satan/Assets/Scripts/Bomb/BombSpawn.cs
@@ -9,13 +9,12 @@ public class BombSpawn : MonoBehaviour
     public GameObject bombType;
 
     private float period = 0.0f;
-    public int spawnTime = 1;
+    public float spawnTime = 1;
 
 
     // Start is called before the first frame update
     void Start()
     {
-
 
     }
